+++ backend/orchestrator/client.py
@@ -32,6 +32,7 @@ from orchestrator.info_pb2_grpc import InfoServiceStub
 from orchestrator.model_pb2_grpc import ModelServiceStub
 from orchestrator.organization_pb2_grpc import OrganizationServiceStub
 from orchestrator.performance_pb2_grpc import PerformanceServiceStub
+from orchestrator.resources import TAG_KEY
 from orchestrator.resources import Algo
 from orchestrator.resources import ComputePlan
 from orchestrator.resources import ComputeTask
@@ -47,7 +48,7 @@ GRPC_RETRYABLE_ERRORS = [
 
 
 def add_tag_from_metadata(task: dict) -> None:
-    task["tag"] = task["metadata"].pop("__tag__", "")
+    task["tag"] = task["metadata"].pop(TAG_KEY, "")
 
 
 def grpc_retry(func):
@@ -355,20 +356,6 @@ class OrchestratorClient:
     def register_models(self, args):
         data = self._model_client.RegisterModels(model_pb2.RegisterModelsParam(**args), metadata=self._metadata)
         return MessageToDict(data, **CONVERT_SETTINGS).get("models", [])
-
-    @grpc_retry
-    def can_disable_model(self, model_key):
-        data = self._model_client.CanDisableModel(
-            model_pb2.CanDisableModelParam(model_key=model_key), metadata=self._metadata
-        )
-        return MessageToDict(data, **CONVERT_SETTINGS).get("can_disable")
-
-    @grpc_retry
-    def disable_model(self, model_key):
-        data = self._model_client.DisableModel(
-            model_pb2.DisableModelParam(model_key=model_key), metadata=self._metadata
-        )
-        return MessageToDict(data, **CONVERT_SETTINGS)
 
     @grpc_retry
     def register_performance(self, args):